+++ src/main/java/fm/last/commons/test/extensions/ClassDataFolderExtension.java
@@ -5,17 +5,19 @@ import java.io.File;
 import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 
-public class ClassDataFolderExtension extends DataFolder implements BeforeEachCallback {
+public class ClassDataFolderExtension extends BaseDataFolder implements BeforeEachCallback {
 
-  private final File parent;
+  private final File parent = new File("src" + File.separator + "test" + File.separator + "data");
+  private File folder;
 
-  public ClassDataFolderExtension(){
-    parent = new File("src" + File.separator + "test" + File.separator + "data");
+  @Override
+  public void beforeEach(ExtensionContext extensionContext) {
+    Class<?> targetClass = extensionContext.getRequiredTestClass();
+    folder = new File(parent, targetClass.getName().replaceAll(PACKAGE_DELIMITER_PATTERN, FILE_SEPARATOR_REPLACEMENT));
   }
 
   @Override
-  public void beforeEach(ExtensionContext extensionContext) throws Exception {
-    Class<?> targetClass = extensionContext.getRequiredTestClass();
-    folder = new File(parent, targetClass.getName().replaceAll(PACKAGE_DELIMITER_PATTERN, FILE_SEPARATOR_REPLACEMENT));
+  public File getDataFolder() {
+    return folder;
   }
 }