+++ src/main/java/fm/last/commons/test/extensions/MethodDataFolderExtension.java
@@ -6,23 +6,25 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 
-public class MethodDataFolderExtension extends DataFolder implements BeforeEachCallback {
+public class MethodDataFolderExtension extends BaseDataFolder implements BeforeEachCallback {
 
-  private final File parent;
+  private final File parent = new File("src" + File.separator + "test" + File.separator + "data");
+  private File folder;
 
-  public MethodDataFolderExtension(){
-    parent = new File("src" + File.separator + "test" + File.separator + "data");
+  @Override
+  public void beforeEach(ExtensionContext extensionContext) {
+    Test testAnnotation = extensionContext.getRequiredTestMethod().getAnnotation(Test.class);
+    if(testAnnotation != null) {
+      String methodName = extensionContext.getRequiredTestMethod().getName();
+      Class<?> targetClass = extensionContext.getRequiredTestClass();
+      folder = new File(parent, targetClass.getName().replaceAll(PACKAGE_DELIMITER_PATTERN,
+          FILE_SEPARATOR_REPLACEMENT)
+          + File.separator + methodName);
+    }
   }
 
   @Override
-  public void beforeEach(ExtensionContext extensionContext) throws Exception {
-    Test testAnnotation = extensionContext.getRequiredTestMethod().getAnnotation(Test.class);
-    if(testAnnotation != null) {
-      final String methodName = extensionContext.getRequiredTestMethod().getName();
-      final Class<?> targetClass = extensionContext.getRequiredTestClass();
-          folder = new File(parent, targetClass.getName().replaceAll(PACKAGE_DELIMITER_PATTERN,
-              FILE_SEPARATOR_REPLACEMENT)
-              + File.separator + methodName);
-    }
+  public File getDataFolder() {
+    return folder;
   }
 }