+++ src/main/java/fm/last/commons/test/extensions/RootSqlFolderExtension.java
@@ -2,31 +2,24 @@ package fm.last.commons.test.extensions;
 
 import java.io.File;
 
-import org.junit.jupiter.api.extension.BeforeEachCallback;
-import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.api.extension.Extension;
 
-public class RootSqlFolderExtension extends DataFolder implements BeforeEachCallback {
+import fm.last.commons.test.core.DataFolderUtils;
 
-  public RootSqlFolderExtension(){
+class RootSqlFolderExtension extends BaseDataFolder implements Extension {
+
+  private File folder;
+
+  RootSqlFolderExtension() {
     this(new String[] {});
   }
 
-  public RootSqlFolderExtension(String... children) {
-    StringBuilder path = new StringBuilder();
-    path.append("src");
-    path.append(File.separator);
-    path.append("test");
-    path.append(File.separator);
-    path.append("sql");
-    for (String child : children) {
-      path.append(File.separator);
-      path.append(child);
-    }
-    folder = new File(path.toString());
+  RootSqlFolderExtension(String... children) {
+    folder = DataFolderUtils.buildRootSqlDataFolder(children);
   }
 
   @Override
-  public void beforeEach(ExtensionContext extensionContext) throws Exception {
-    //Register extension
+  public File getDataFolder() {
+    return folder;
   }
 }