+++ src/main/java/fm/last/commons/test/file/DataFolder.java
@@ -23,20 +23,22 @@ import java.util.regex.Pattern;
 
 import org.junit.rules.TestRule;
 
-public abstract class DataFolder implements TestRule, BaseDataFolder {
+import fm.last.commons.test.core.DataFolderUtils;
+
+/**
+ * NOTE: This class shares the same name as the interface due to backwards compatibility issues
+ * that were introduced when moving this project to JUnit5.
+ */
+public abstract class DataFolder implements TestRule, fm.last.commons.test.core.DataFolder {
 
   static final String FILE_SEPARATOR_REPLACEMENT = Matcher.quoteReplacement(File.separator);
   static final String PACKAGE_DELIMITER_PATTERN = Pattern.quote(".");
 
-  File folder;
-  DataFolderCore core = new DataFolderCore();
-
-  DataFolder() {
-  }
+  public abstract File getDataFolder();
 
   @Override
   public File getFolder() throws IOException {
-    return core.getFolder(folder);
+    return DataFolderUtils.getFolder(getDataFolder());
   }
 
   @Override
@@ -46,12 +48,12 @@ public abstract class DataFolder implements TestRule, BaseDataFolder {
 
   @Override
   public String getAbsolutePath(String relativePath) throws IOException {
-    return core.internalGetAbsolutePath(relativePath, folder);
+    return DataFolderUtils.getAbsolutePath(relativePath, getDataFolder());
   }
 
   @Override
   public URI getUri(String relativePath) throws IOException {
-    return core.getUri(relativePath, folder);
+    return DataFolderUtils.getUri(relativePath, getDataFolder());
   }
 
 }