+++ src/test/java/fm/last/commons/test/extensions/SuffixedRootDataFolderExtensionTest.java
@@ -1,6 +1,6 @@
 package fm.last.commons.test.extensions;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.Assert.assertTrue;
 
 import java.io.File;
 import java.io.IOException;
@@ -10,15 +10,18 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 
 import fm.last.commons.test.core.DataFolder;
 
-public class ClassDataFolderExtensionTest {
+public class SuffixedRootDataFolderExtensionTest {
 
   @RegisterExtension
-  DataFolder folder = new ClassDataFolderExtension();
+  DataFolder folder = new RootDataFolderExtension("suffix");
 
   @Test
-  public void initialised() throws IOException {
+  public void integration() throws IOException {
     File actualFolder = folder.getFolder();
     assertTrue(actualFolder.exists());
     assertTrue(actualFolder.canRead());
+    assertTrue(actualFolder.getAbsolutePath().endsWith(
+        "src" + File.separator + "test" + File.separator + "data" + File.separator + "suffix"));
   }
+
 }